Add bridge data to fills endpoint
